chatbot/requirements.txt
flask==3.0.2
python-dotenv==1.0.1
openai==1.59.3
pydantic==2.9.2
flask-cors==4.0.1
